% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daterangepicker.R
\name{perioddaterangeOptions}
\alias{perioddaterangeOptions}
\title{perioddaterangeOptions}
\usage{
perioddaterangeOptions(
  period = c("day", "week", "month", "quarter", "year"),
  periods = NULL,
  firstDayOfWeek = 0,
  single = FALSE,
  orientation = c("right", "left"),
  opened = FALSE,
  expanded = FALSE,
  standalone = FALSE,
  hideWeekdays = FALSE,
  anchorElement = NULL,
  parentElement = NULL,
  forceUpdate = FALSE,
  timezone = "utc",
  locale = list(applyButtonTitle = "Apply", cancelButtonTitle = "Cancel", endLabel =
    "End", inputFormat = "L", startLabel = "Start", dayLabel = "Day", weekLabel = "Week",
    monthLabel = "Month", quarterLabel = "Quarter", yearLabel = "Year")
)
}
\arguments{
\item{period}{This parameter sets the initial value for period. Must be one
of \code{c('day','week','month','quarter','year')}}

\item{periods}{Array of available periods. Period selector disappears if
only one period specified.}

\item{firstDayOfWeek}{Sets first day of the week. 0 is Sunday, 1 is Monday.
In case you were wondering, 4 is Thursday.}

\item{single}{Should only a single date be returned. Default is \code{FALSE}}

\item{orientation}{Sets the side to which daterangepicker opens. Must be one
of \code{c('left','right')}.}

\item{opened}{By default, daterangepicker is hidden and you need to click the
\code{anchorElement} to open it. This option allows you to make it opened
on initialization. Default is \code{FALSE}}

\item{expanded}{By default, when you open daterangepicker you only see
predefined ranges. This option allows you to make it expanded on
initialization. Default is \code{FALSE}}

\item{standalone}{Set standalone to true to append daterangepicker to
\code{anchorElement}. Default is \code{FALSE}}

\item{hideWeekdays}{Set to \code{TRUE} to hide week days in day & week modes.}

\item{anchorElement}{Allows you to set anchor element for daterangepicker.}

\item{parentElement}{Allows you to set parent element for daterangepicker.}

\item{forceUpdate}{Immediately invokes callback after constructing
daterangepicker.}

\item{locale}{A list with the Date format and Button labels.}
}
\description{
Update the perioddaterangepicker
}
\seealso{
\href{https://sensortower.github.io/daterangepicker/docs}{https://sensortower.github.io/}

Other perioddaterangepicker Functions: 
\code{\link{perioddaterangepicker}()}
}
\concept{perioddaterangepicker Functions}
