% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daterangepicker.R
\name{perioddaterangepicker}
\alias{perioddaterangepicker}
\title{perioddaterangepicker}
\usage{
perioddaterangepicker(
  inputId = NULL,
  label = "Select a Date",
  start = NULL,
  end = NULL,
  min = NULL,
  max = NULL,
  ranges = NULL,
  language = "en",
  timezone = "utc",
  style = "width:100\%;border-radius:4px;text-align:center;",
  class = NULL,
  icon = NULL,
  options = perioddaterangeOptions()
)
}
\arguments{
\item{inputId}{The input ID}

\item{label}{The label for the control, or NULL for no label.}

\item{start}{The beginning date of the initially selected. Must be a Date /
POSIXt or string. If NULL will default to the current day.}

\item{end}{The end date of the initially selected date range. Must be a Date
/ POSIXt or string. If NULL will default to the current day.}

\item{min}{The earliest date a user may select. Must be a Date or string}

\item{max}{The latest date a user may select. Must be a Date or string}

\item{ranges}{Set predefined date ranges the user can select from. Each key
is the label for the range, and its value an array with two dates
representing the bounds of the range.}

\item{language}{The language used for month and day names. Default is "en".
See the \href{https://momentjs.com/}{Multiple Locale Support} for a list of
other valid values.}

\item{timezone}{Sets the time zone. If you want to use user's computer
time zone, pass \code{NULL}. By default, it's `utc`}

\item{style}{Add CSS-styles to the input.}

\item{class}{Custom class}

\item{icon}{Icon to display next to the label.}

\item{options}{List of further options. See
\code{\link{perioddaterangeOptions}}}
}
\description{
The Date Range Picker pops up two calendars for selecting dates, times, or
predefined ranges like "Yesterday", "Last 30 Days", etc.
}
\seealso{
\href{https://sensortower.github.io/daterangepicker/}{https://sensortower.github.io/}

Other perioddaterangepicker Functions: 
\code{\link{perioddaterangeOptions}()}
}
\concept{perioddaterangepicker Functions}
